[C++] Support C++ object copies
Augment the C++ generator to emit C++ copy/move constructors and
assignment operators. This is enabled by default when the C++ standard
is C++11 or later. These additional functions are only emitted for
objects that need it.

The copy ctor and assignment operator are declared in the object table
type and are defined as inline functions after table declarations.

When copy/move ctors and assignment operators are declared, a
user-defined explicitly-defaulted default ctor is also emitted.

fixes #5783